pyproject.toml
[tool.poetry]
name = "collection"
version = "0.1.0"
description = "Приложение групповых денежных сборов"
authors = ["Kozin Artem <kaf-artem@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.0"
Django = "5.0.2"
djangorestframework = "3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.1"
djoser = "^2.2.2"
gunicorn = "^21.2.0"
django-filter = "^23.5"
drf-extra-fields = "^3.7.0"
django-redis-cache = "^3.0.1"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "1.7.0"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79
remove_redundant_aliases = true
skip_glob = ["**/migrations/*.py"]
src_paths = ["src/backend/"]

[tool.black]
line-length = 79
target-version = ['py312']
include = "src/backend/"
extend-exclude = """
/(
  | migrations
)/
"""

[tool.mypy]
exclude = "migrations"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.prod"
